Adding JWT Claims

+++ section11/resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from models.item import ItemModel
 
 class Item(Resource):
@@ -17,7 +17,7 @@ class Item(Resource):
         help="Every item needs a store_id."
                         )
 
-    @jwt_required
+    @jwt_required()
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -39,15 +39,19 @@ class Item(Resource):
 
         return item.json(), 201
 
-    @jwt_required
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'message': 'Admin privilege required.'}, 401
+    
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
             return {'message': 'Item deleted.'}
         return {'message': 'Item not found.'}, 404
 
-    @jwt_required
+    @jwt_required()
     def put(self, name):
         data = Item.parser.parse_args()
         item = ItemModel.find_by_name(name)
